Settled on implementation for object loading cache
Also tried with memoized hashes in a separate table but not better,
slightly worse, probably due to less locality.
