--- src/main/java/grondag/adversity/library/cache/objectKey/IObjectLoadingCache.java
@@ -1,10 +1,0 @@
-package grondag.adversity.library.cache.objectKey;
-
-public interface IObjectLoadingCache<K, V>
-{
-
-    void clear();
-
-    V get(K key);
-
-}
