+++ src/main/java/grondag/adversity/library/cache/objectKey/ObjectSimpleCacheLoader.java
@@ -1,6 +1,8 @@
 package grondag.adversity.library.cache.objectKey;
 
-public interface ObjectSimpleCacheLoader<K,V>
+public interface ObjectSimpleCacheLoader<K, V>
 {
     abstract public V load(K key);
+    
+    abstract ObjectSimpleCacheLoader<K, V> createNew();
 }
